GooglePlaceApi split
