+++ src/main/java/com/example/demo/controller/voteStore/VoteStoreController.java
@@ -1,6 +1,7 @@
 package com.example.demo.controller.voteStore;
 
 import com.example.demo.service.voteStore.VoteStoreService;
+import com.example.demo.utils.GooglePlaceApi;
 import com.fasterxml.jackson.databind.JsonNode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -16,6 +17,9 @@ public class VoteStoreController {
     @Autowired
     VoteStoreService voteStoreService;
 
+    @Autowired
+    GooglePlaceApi googlePlaceApi;
+
     @RequestMapping("/voteStore")
     public String voteSotre() {
         return "voteStore";
@@ -26,15 +30,9 @@ public class VoteStoreController {
     public List<JsonNode> storeInfo(@PathVariable("date") String date) {
         System.out.println(date);
         List<String> store = voteStoreService.getStoreId(date);
-        for (int i = 0; i < store.size(); i++) {
-            System.out.println(store.get(i));
-        }
-        voteStoreService.init(store);
-        List<JsonNode> result = voteStoreService.getData();
 
-        for (JsonNode node : result) {
-            System.out.println(node);
-        }
+        googlePlaceApi.init(store);
+        List<JsonNode> result = googlePlaceApi.getData();
 
         return result;
     }
